code refactory is done
